+++ src/main/java/info/macias/kaconf/sources/JavaUtilPropertySource.java
@@ -15,10 +15,9 @@
 */
 package info.macias.kaconf.sources;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.io.InputStream;
+import info.macias.kaconf.ConfiguratorException;
+
+import java.io.*;
 import java.util.Properties;
 
 /**
@@ -34,32 +33,17 @@ public class JavaUtilPropertySource extends AbstractPropertySource {
      * the user does not have permissions), no exceptions will be thrown and the object will be
      * instantiated anyway. However, the {@link JavaUtilPropertySource#isAvailable()} method will
      * return <code>false</code>.</p>
+     *
      * @param filePath The path to reach the Properties file.
+     * @deprecated Use {@link #from(String)} instead. This constructor will be removed in version 0.9.0.
      */
     public JavaUtilPropertySource(String filePath) {
-        this(new File(filePath));
-    }
-
-    /**
-     * <p>Instantiates the class by loading the File specified as argument.</p>
-     * <p>If the {@link java.util.Properties} cannot be loaded (e.g. because the file does not exist or
-     * the user does not have permissions), no exceptions will be thrown and the object will be
-     * instantiated anyway. However, the {@link JavaUtilPropertySource#isAvailable()} method will
-     * return <code>false</code>.</p>
-     * @param filePath The path to reach the Properties file.
-     */
-    public JavaUtilPropertySource(File file) {
-        if (file == null || !file.isFile())
-            throw new IllegalArgumentException("'" + file + "' is not a valid file.");
-        try(FileInputStream fis = new FileInputStream(file)) {
-            Properties props = new Properties();
-            props.load(fis);
-            properties = props;
+        try(FileInputStream fis = new FileInputStream(filePath)) {
+            properties.load(fis);
         } catch (IOException e) {
             properties = null;
         }
     }
-
     /**
      * <p>Instantiates the class by loading the {@link java.util.Properties} object that is
      * reachable by the {@link InputStream} specified as argument</p>
@@ -67,7 +51,9 @@ public class JavaUtilPropertySource extends AbstractPropertySource {
      * the user does not have permissions), no exceptions will be thrown and the object will be
      * instantiated anyway. However, the {@link JavaUtilPropertySource#isAvailable()} method will
      * return <code>false</code>.</p>
+     *
      * @param is The {@link InputStream} to access the Properties file
+     * @deprecated Use {@link #from(InputStream)} instead. This constructor will be removed in version 0.9.0.
      */
     public JavaUtilPropertySource(InputStream is) {
         try {
@@ -80,8 +66,61 @@ public class JavaUtilPropertySource extends AbstractPropertySource {
     }
 
     /**
+     * <p>Creates a properties source by loading the properties file with the path that is specified as argument.</p>
+     *
+     * @param filePath The path to reach the Properties file.
+     * @throws ConfiguratorException if there is something wrong when loading the properties.
+     */
+    public static JavaUtilPropertySource from(String filePath) {
+        File file = new File(filePath);
+        return from(file);
+    }
+
+    /**
+     * <p>Creates a properties source by loading the File specified as argument.</p>
+     *
+     * @param file The {@link File} object to reach the Properties file.
+     * @throws ConfiguratorException if there is something wrong when loading the properties.
+     */
+    public static JavaUtilPropertySource from(File file) {
+        if (file == null)
+            throw new IllegalArgumentException("'" + file + "' is not a valid file.");
+        Properties properties;
+        try (FileInputStream fis = new FileInputStream(file)) {
+            properties = new Properties();
+            properties.load(fis);
+        } catch (IOException e) {
+            throw new ConfiguratorException(e);
+        }
+        return new JavaUtilPropertySource(properties);
+    }
+
+    /**
+     * <p>Creates a properties source by loading the {@link java.util.Properties} object that is
+     * reachable by the {@link InputStream} specified as argument</p>
+     *
+     * @param is The {@link InputStream} to access the Properties file
+     * @throws ConfiguratorException if there is something wrong when loading the properties.
+     */
+    public static JavaUtilPropertySource from(InputStream is) {
+        if (is == null) {
+            throw new IllegalArgumentException("The provided Input Stream can't be null");
+        }
+        Properties properties;
+        try {
+            properties = new Properties();
+            properties.load(is);
+        } catch (NullPointerException | IOException e) {
+            properties = null;
+        }
+        return new JavaUtilPropertySource(properties);
+    }
+
+
+    /**
      * <p>Instantiates the class to handle the {@link java.util.Properties} object passed as an
      * argument</p>
+     *
      * @param properties The properties that will be handled by the instantiated object
      */
     public JavaUtilPropertySource(Properties properties) {
@@ -93,7 +132,7 @@ public class JavaUtilPropertySource extends AbstractPropertySource {
      * loaded. Otherwise, returns <code>false</code>.
      *
      * @return <code>true</code> if the {@link java.util.Properties} object has been correctly
-     *          loaded. <code>false</code> otherwise
+     * loaded. <code>false</code> otherwise
      */
     @Override
     public boolean isAvailable() {
@@ -106,6 +145,9 @@ public class JavaUtilPropertySource extends AbstractPropertySource {
      */
     @Override
     protected String get(String key) {
+        if (!isAvailable()) {
+            return null;
+        }
         return properties.getProperty(key);
     }
 }