+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v0.8.4
 * Fixed `NullPointerException` error on `JavaUtilPropertySource`.
 * Added `URL` to the list of supported configurable types.
+* Deprecated file-based constructors for `JavaUtilPropertySource` (use
+  `JavaUtilPropertySource.from(*)` static method instead)
 
 ## v0.8.2
 * First public version.